mysqlにuserを追加, redisは未

+++ domain/entity/user.go
@@ -4,8 +4,8 @@ type User struct {
 	ID           int64
 	Name         string
 	SubmitID     string
-	Sex          int64
-	Year         int64
+	Sex          int
+	Year         int
 	PasswordHash string
 	Clothes      []Clothe
 }
